Rewrite Corpus Downloading
Handle the confusion about the PAUs REST API being able to take either
a PAU ID or a TREC ID as input.

Always read IDs from the returned PAU structure. Walk the entire "hits" list.

This requires that the document IDs and PAUs be written to checkpoint files
at the same time.