Extract Corpus from TREC Files
Extract the corpus from a directory of TREC XML files instead of
downloading it over the wire.

TREC XML is malformed (contains &, <, and > characters in text)
so parse it with Beautiful Soup.

Checkpoint this process because it can take a long time.
Checkpointing logic refactored out. Should eventually have the
other checkpoiinted processes use this refactoring.