Handle Bad XMGR Error When Downloading Truth File
If you use an invalid URL to download the truth file from XMGR it returns
HTTP 200 and puts the error message in text. This results in an unhelpful
error message about being unable to decode JSON.

Catch this case and throw a more useful error about a bad URL.

I had already handled this for a bad URL when downloading the corpus.