Augment Corpus
Augment the corpus with answers pulled from the usage logs and PAUs
downloaded from XMGR whose IDs are in the truth file.

You would expect all the answers returned by the system to be in the
corpus, but this is not the case. The  'themis xmgr validate-answers'
and 'themis xmgr validate-truth' commands shows which answers are
missing from the corpus.